+++ aksel.nav.no/website/components/sanity-modules/cards/RelatertInnhold.tsx
@@ -27,7 +27,7 @@ const RelatertInnhold = ({
       className={cl(
         "relatedCard",
         "mb-8 max-w-4xl gap-4",
-        "xs:grid-cols-2 grid w-full"
+        "grid w-full sm:grid-cols-2"
       )}
     >
       {node.lenker.map((x) => (
@@ -42,7 +42,7 @@ const RelatertInnhold = ({
               e.currentTarget.getAttribute("href")
             )
           }
-          className="algolia-ignore-index shadow-small focus-visible:shadow-focus bg-surface-default ring-border-subtle group grid rounded border-2 border-transparent px-4 py-3 ring-1 focus:outline-none"
+          className="toc-ignore shadow-small focus-visible:shadow-focus bg-surface-default ring-border-subtle group grid rounded border-2 border-transparent px-4 py-3 ring-1 focus:outline-none"
         >
           <Heading
             as="span"